[expo-notifications] Split ExpoNotificationsService
